requirements.txt
# OptiFIRE Dependencies
# Python 3.11+

# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Async & Concurrency
aiohttp==3.9.1
aiosqlite==0.19.0
httpx==0.26.0

# Database
sqlalchemy==2.0.25
duckdb==0.9.2
pyarrow==14.0.2

# Scheduling
apscheduler==3.10.4

# Auth
pyjwt==2.8.0
pyotp==2.9.0

# Data Science
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4

# ML/AI (lazy load where possible)
lightgbm==4.2.0
scikit-learn==1.3.2
optuna==3.5.0
sentence-transformers==2.2.2
onnxruntime==1.16.3

# Risk & Finance
statsmodels==0.14.1

# Monitoring
psutil==5.9.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Templates
jinja2==3.1.3

# YAML
pyyaml==6.0.1

# Misc
python-dotenv==1.0.0
